Create Transport.CounterCache module
